+++ lib/Components/VariableList/__test__/VariableList.spec.jsx
@@ -13,8 +13,8 @@ import { ScopeExpandProvider } from '../../../Context/ScopeExpandContext';
 
 const defaultFilter = {
   search: '',
-  filterType: 'all',
   selectedElements: [],
+  writtenOnly: false
 };
 
 let wrapper;
@@ -92,9 +92,11 @@ describe('VariableList', () => {
           variables: [ 'OuterSubprocessInputVariable1' ],
         }
       ]);
+
     }));
 
   });
+
 
   describe('given ServiceTask is selected', () => {