+++ lib/index.jsx
@@ -19,8 +19,8 @@ export default function(props) {
 
   const search = useState({
     search: '',
-    filterType: 'all',
     selectedElements: [],
+    writtenOnly: false
   });
 
   return (